Legger til filer

+++ src/main/java/service/UserService.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class UserService {
 
     @Inject
-    private UserDAO userDAO;
+    public UserDAO userDAO;
 
 
     public List<User> getAllUsers(){
@@ -24,12 +24,23 @@ public class UserService {
     }
 
     public boolean deleteUserByID(int id){
+        if(id < 0) return false;
+
         User user = userDAO.getUser(id);
+
+        if(user == null){
+            return false;
+        }
+
         return userDAO.deleteUser(user);
     }
 
     public boolean updateUser(User user){
         return userDAO.updateUser(user);
+    }
+
+    public User getUserByID(int id){
+        return userDAO.getUser(id);
     }
 
 }
